migração para postgres Ufs, Municipios e Bairros ok

+++ src/database/migrations/1644262447158-CreateUfs.ts
@@ -23,7 +23,7 @@ export class CreateUfs1644262447158 implements MigrationInterface {
             },
             {
               name: "STATUS",
-              type: "integer"
+              type: "numeric"
             }
           ]
         })
